test(node): remove unnecessary footer test
