+++ packages/rolldown/tests/fixtures/function/footer/string/_config.ts
@@ -1,0 +1,15 @@
+import { expect } from 'vitest'
+import { defineTest } from 'rolldown-tests'
+
+const footerTxt = '// footer test\n'
+
+export default defineTest({
+  config: {
+    output: {
+      footer:footerTxt
+    },
+  },
+  afterTest: (output) => {
+    expect(output.output[0].code.endsWith(footerTxt)).toBe(true)
+  },
+})
